move `ProgramEnvironment` into the chalk crate
