+++ src/coherence/orphan.rs
@@ -1,11 +1,11 @@
 use std::sync::Arc;
 
 use super::CoherenceError;
+use crate::program_environment::ProgramEnvironment;
 use crate::rust_ir::*;
 use chalk_ir::cast::*;
 use chalk_ir::*;
 use chalk_solve::ext::*;
-use chalk_solve::program_environment::ProgramEnvironment;
 use chalk_solve::solve::SolverChoice;
 use failure::Fallible;
