+++ src/coherence/solve.rs
@@ -1,15 +1,14 @@
-use std::sync::Arc;
-
 use super::CoherenceError;
+use crate::program_environment::ProgramEnvironment;
 use crate::rust_ir::*;
 use chalk_ir::cast::*;
 use chalk_ir::fold::shift::Shift;
 use chalk_ir::*;
 use chalk_solve::ext::*;
-use chalk_solve::program_environment::ProgramEnvironment;
 use chalk_solve::solve::{Solution, SolverChoice};
 use failure::Fallible;
 use itertools::Itertools;
+use std::sync::Arc;
 
 struct DisjointSolver {
     env: Arc<ProgramEnvironment>,
